adding and moving tests to main

+++ src/Djikstras.cpp
@@ -15,7 +15,7 @@ std::vector<int> Djikstras::finalvector(int currentVertex, std::vector<int> pare
    return final_vect;
 }
 
-std::vector<int> Djikstras::dijkstra(std::vector<std::vector<int>> adjacencyMatrix, int src, int destination) {
+std::vector<int> Djikstras::dijkstra(std::vector<std::vector<double>> adjacencyMatrix, int src, int destination) {
    int vertices = adjacencyMatrix[0].size();
    vector<int> distances(vertices);
    vector<bool> added(vertices);
@@ -31,13 +31,16 @@ std::vector<int> Djikstras::dijkstra(std::vector<std::vector<int>> adjacencyMatr
 
    for (int i = 1; i < vertices; i++) {
       int neighbor = -1;
-      int shortest = INT_MAX;
+      double shortest = INT_MAX;
       for (int j = 0; j < vertices; j++) {
+         //std::cout << "j = " << j << " , added[j]=" << added[j] << " , dist[j]=" << distances[j] << std::endl;
          if (!added[j] && distances[j] < shortest) {
             neighbor = j;
             shortest = distances[j];
          }
       }
+      
+      //std::cout << "neighbor = " << neighbor << std::endl;
       added[neighbor] = true;
       for (int j = 0; j < vertices; j++) {
          int dist = adjacencyMatrix[neighbor][j];
@@ -52,4 +55,42 @@ std::vector<int> Djikstras::dijkstra(std::vector<std::vector<int>> adjacencyMatr
    final_vect = finalvector(destination, prev);
    return final_vect;
 }
+
+// std::vector<int> Djikstras::dijkstra(std::vector<std::vector<double>> adjacencyMatrix, int src, int destination) {
+//    int vertices = adjacencyMatrix[0].size();
+//    vector<int> distances(vertices);
+//    vector<bool> added(vertices);
+
+//    for (int i = 0; i < vertices; i++) {
+//       distances[i] = INT_MAX;
+//       added[i] = false;
+//    }
+
+//    distances[src] = 0;
+//    std::vector<int> prev(vertices);
+//    prev[src] = -1;
+
+//    for (int i = 1; i < vertices; i++) {
+//       int neighbor = -1;
+//       int shortest = INT_MAX;
+//       for (int j = 0; j < vertices; j++) {
+//          if (!added[j] && distances[j] < shortest) {
+//             neighbor = j;
+//             shortest = distances[j];
+//          }
+//       }
+//       added[neighbor] = true;
+//       for (int j = 0; j < vertices; j++) {
+//          int dist = adjacencyMatrix[neighbor][j];
+//          if (dist > 0 && ((shortest + dist) < distances[j])) {
+//             prev[j] = neighbor;
+//             distances[j] = shortest + dist;
+//          }
+//       }
+//    }
+
+//    std::vector<int> final_vect;
+//    final_vect = finalvector(destination, prev);
+//    return final_vect;
+// }
